initial compilation of existing and new classes (std=c++14) OK

+++ bobcat/csv4180/read.cc
@@ -5,7 +5,7 @@ size_t CSV4180::read(istream &in, size_t nLines)
     size_t idx = 0;
 
     if (nLines == 0)
-        nLines = static_cast<size_t>(~0U);
+        nLines = ~0;
 
     for (; idx != nLines; ++idx)        // read all required lines or stop
     {                                   // once reading fails.
@@ -13,9 +13,8 @@ size_t CSV4180::read(istream &in, size_t nLines)
             break;
     }
 
-    if (nLines == static_cast<size_t>(~0U)) // at undetermined # lines:
+    if (nLines == static_cast<size_t>(~0))  // at undetermined # lines:
         in.clear();                         //  reset `in' as it may contain
                                             //  additional info
-
     return d_data.size();
 }
